Refactored permissions middleware to make it easier to use." "
